Condense all terraforming code into a single file
